# frontend/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Local Doujin Studio",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/projects/{project_id}/story-sessions": {
      "post": {
        "summary": "Create Story Session",
        "operationId": "create_story_session_api_projects__project_id__story_sessions_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "revision",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Revision"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StorySessionCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectMutationResponse_StorySessionResponse_"
                }
              }
            }
          },
          "409": {
            "description": "revision競合。最新のprojectを同梱する",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectRevisionConflictResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List Story Sessions",
        "operationId": "list_story_sessions_api_projects__project_id__story_sessions_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StorySessionSummary"
                  },
                  "title": "Response List Story Sessions Api Projects  Project Id  Story Sessions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/story-sessions/{session_id}": {
      "get": {
        "summary": "Get Story Session",
        "operationId": "get_story_session_api_story_sessions__session_id__get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StorySessionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/story-sessions/{session_id}/stages/{stage}/generate": {
      "post": {
        "summary": "Generate Story Stage",
        "operationId": "generate_story_stage_api_story_sessions__session_id__stages__stage__generate_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          },
          {
            "name": "stage",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Stage"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/StageGenerateRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StorySessionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/story-sessions/{session_id}/stages/{stage}": {
      "put": {
        "summary": "Update Story Stage",
        "operationId": "update_story_stage_api_story_sessions__session_id__stages__stage__put",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          },
          {
            "name": "stage",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Stage"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StageUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StorySessionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/story-sessions/{session_id}/stages/{stage}/approve": {
      "post": {
        "summary": "Approve Story Stage",
        "operationId": "approve_story_stage_api_story_sessions__session_id__stages__stage__approve_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          },
          {
            "name": "stage",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Stage"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StorySessionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/story-sessions/{session_id}/apply": {
      "post": {
        "summary": "Apply Story Session",
        "operationId": "apply_story_session_api_story_sessions__session_id__apply_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          },
          {
            "name": "revision",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Revision"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectMutationResponse_EmptyMutationResult_"
                }
              }
            }
          },
          "409": {
            "description": "revision競合。最新のprojectを同梱する",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectRevisionConflictResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/revisions": {
      "get": {
        "summary": "List Project Revisions",
        "operationId": "list_project_revisions_api_projects__project_id__revisions_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProjectRevisionResponse"
                  },
                  "title": "Response List Project Revisions Api Projects  Project Id  Revisions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/revisions/{revision_id}/restore": {
      "post": {
        "summary": "Restore Project Revision",
        "operationId": "restore_project_revision_api_projects__project_id__revisions__revision_id__restore_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "revision_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Revision Id"
            }
          },
          {
            "name": "revision",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Revision"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectMutationResponse_EmptyMutationResult_"
                }
              }
            }
          },
          "409": {
            "description": "revision競合。最新のprojectを同梱する",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectRevisionConflictResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "EmptyMutationResult": {
        "properties": {},
        "type": "object",
        "title": "EmptyMutationResult",
        "description": "操作固有情報がないmutation用の空result。"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MangaProject": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "work_name": {
            "type": "string",
            "title": "Work Name",
            "default": ""
          },
          "premise": {
            "type": "string",
            "title": "Premise",
            "default": ""
          },
          "target_pages": {
            "type": "integer",
            "title": "Target Pages",
            "default": 4
          },
          "reading_direction": {
            "type": "string",
            "enum": [
              "rtl",
              "ltr"
            ],
            "title": "Reading Direction",
            "default": "rtl"
          },
          "typography": {
            "$ref": "#/components/schemas/TypographySettings"
          },
          "page_layout": {
            "$ref": "#/components/schemas/PageLayoutSettings"
          },
          "common_positive_prompt": {
            "type": "string",
            "title": "Common Positive Prompt",
            "default": ""
          },
          "common_negative_prompt": {
            "type": "string",
            "title": "Common Negative Prompt",
            "default": ""
          },
          "characters": {
            "items": {
              "$ref": "#/components/schemas/Character"
            },
            "type": "array",
            "title": "Characters"
          },
          "locations": {
            "items": {
              "$ref": "#/components/schemas/LocationProfile"
            },
            "type": "array",
            "title": "Locations"
          },
          "workflow_presets": {
            "items": {
              "$ref": "#/components/schemas/WorkflowPreset"
            },
            "type": "array",
            "title": "Workflow Presets"
          },
          "active_workflow_preset_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Workflow Preset Id"
          },
          "pages": {
            "items": {
              "$ref": "#/components/schemas/Page"
            },
            "type": "array",
            "title": "Pages"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "title"
        ],
        "title": "MangaProject"
      },
...
      "ProjectDetail": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "work_name": {
            "type": "string",
            "title": "Work Name"
          },
          "revision": {
            "type": "integer",
            "title": "Revision",
            "default": 0
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "manga_json": {
            "$ref": "#/components/schemas/MangaProject"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "work_name",
          "created_at",
          "updated_at",
          "manga_json"
        ],
        "title": "ProjectDetail"
      },
...
      "ProjectMutationResponse_EmptyMutationResult_": {
        "properties": {
          "project": {
            "$ref": "#/components/schemas/ProjectDetail"
          },
          "result": {
            "$ref": "#/components/schemas/EmptyMutationResult"
          }
        },
        "type": "object",
        "required": [
          "project",
          "result"
        ],
        "title": "ProjectMutationResponse[EmptyMutationResult]"
      },
...
      "ProjectMutationResponse_StorySessionResponse_": {
        "properties": {
          "project": {
            "$ref": "#/components/schemas/ProjectDetail"
          },
          "result": {
            "$ref": "#/components/schemas/StorySessionResponse"
          }
        },
        "type": "object",
        "required": [
          "project",
          "result"
        ],
        "title": "ProjectMutationResponse[StorySessionResponse]"
      },
...
      "ProjectRevisionConflictResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "code": {
            "type": "string",
            "const": "project_revision_conflict",
            "title": "Code",
            "default": "project_revision_conflict"
          },
          "expected_revision": {
            "type": "integer",
            "title": "Expected Revision"
          },
          "actual_revision": {
            "type": "integer",
            "title": "Actual Revision"
          },
          "project": {
            "$ref": "#/components/schemas/ProjectDetail"
          }
        },
        "type": "object",
        "required": [
          "detail",
          "expected_revision",
          "actual_revision",
          "project"
        ],
        "title": "ProjectRevisionConflictResponse",
        "description": "revision競合時の専用409応答。最新projectを同梱する。"
      },
      "ProjectRevisionResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "label",
          "created_at"
        ],
        "title": "ProjectRevisionResponse"
      },
...
      "StageGenerateRequest": {
        "properties": {
          "instruction": {
            "type": "string",
            "maxLength": 2000,
            "title": "Instruction",
            "default": ""
          }
        },
        "type": "object",
        "title": "StageGenerateRequest"
      },
      "StageUpdateRequest": {
        "properties": {
          "data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "StageUpdateRequest"
      },
      "StorySessionCreate": {
        "properties": {
          "work_name": {
            "type": "string",
            "maxLength": 120,
            "title": "Work Name",
            "default": ""
          },
          "knowledge_work_id": {
            "type": "string",
            "maxLength": 120,
            "title": "Knowledge Work Id",
            "default": ""
          },
          "target_pages": {
            "type": "integer",
            "enum": [
              4,
              8,
              16
            ],
            "title": "Target Pages",
            "default": 4
          },
          "instruction": {
            "type": "string",
            "maxLength": 2000,
            "title": "Instruction",
            "default": ""
          }
        },
        "type": "object",
        "title": "StorySessionCreate"
      },
      "StorySessionResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "work_name": {
            "type": "string",
            "title": "Work Name"
          },
          "target_pages": {
            "type": "integer",
            "title": "Target Pages"
          },
          "instruction": {
            "type": "string",
            "title": "Instruction"
          },
          "stages": {
            "additionalProperties": {
              "$ref": "#/components/schemas/StoryStageState"
            },
            "type": "object",
            "title": "Stages"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "work_name",
          "target_pages",
          "instruction",
          "stages",
          "created_at",
          "updated_at"
        ],
        "title": "StorySessionResponse"
      },
      "StorySessionSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "work_name": {
            "type": "string",
            "title": "Work Name"
          },
          "target_pages": {
            "type": "integer",
            "title": "Target Pages"
          },
          "instruction": {
            "type": "string",
            "title": "Instruction"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "work_name",
          "target_pages",
          "instruction",
          "created_at",
          "updated_at"
        ],
        "title": "StorySessionSummary"
      },
      "StoryStageState": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "empty",
              "draft",
              "approved"
            ],
            "title": "Status",
            "default": "empty"
          },
          "data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "knowledge_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Knowledge Ids"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "title": "StoryStageState"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}